front Start

+++ src/controller/videoController.js
@@ -12,19 +12,26 @@ export const trending = async (req, res) => {
 
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
-    const owner = await User.findById(video.owner);
+    const video = await Video.findById(id).populate("owner");
+    console.log(video);
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
     }
-    return res.render("watch", { pageTitle: video.title, video, owner });
+    return res.render("watch", { pageTitle: video.title, video });
 };
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id },
+    } = req.session;
     const video = await Video.findById(id);
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    console.log(video.owner, _id);
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: video.title, video });
 };
@@ -35,6 +42,9 @@ export const postEdit = async (req, res) => {
     const video = await Video.exists({ _id: id });
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -59,15 +69,17 @@ export const postUpload = async (req, res) => {
         body: { title, description, hashtags },
         file: { path: fileUrl },
     } = req;
-    console.log(req.session);
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             fileUrl,
             owner: _id,
             title,
             description,
             hashtags: Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     } catch (error) {
         return res.status(400).render("upload", {
@@ -78,8 +90,26 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
+    const {
+        session: {
+            user: { _id },
+        },
+    } = req;
     const { id } = req.params;
+    const video = await Video.findById(id).populate("owner");
+    if (!video) {
+        res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner._id) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
+    const { videos: videoList } = video.owner;
+    if (videoList.length > 5) {
+        //todo 남겨진 id목록으로 삭제된 동영상 기능 추가
+        videoList.splice(5, 1);
+        video.owner.save();
+    }
     return res.redirect("/");
 };
